requirements.txt
absl-py==0.13.0
attrs==21.2.0
certifi==2021.5.30
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
mediapipe==0.8.6.2
numpy==1.21.1
opencv-contrib-python==4.5.3.56
Pillow==8.3.1
protobuf==3.17.3
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
wincertstore==0.2
